Add: install interface

+++ environment.sh
@@ -1,15 +1,31 @@
-# Docker Install
-sudo wget -qO- https://get.docker.com/ | sh
-sudo docker rm `sudo docker ps -aq`
-sudo docker rmi hello-world
+echo "========== Docker Install =========="
+echo "Check Docker Installed.."
+if type docker &>/dev/null; then
+    echo "Docker Already Installed."
+else
+    echo "Installing Docker..."
+    sudo wget -qO- https://get.docker.com/ | sh
+    echo "Setting Docker..."
+    sudo docker rm `sudo docker ps -aq`
+    sudo docker rmi hello-world
+    echo "Now, Docker Installed."
+fi
 
-# Nvidia Docker Install
-distribution=$(. /etc/os-release;echo $ID$VERSION_ID)
-curl -s -L https://nvidia.github.io/nvidia-docker/gpgkey | sudo apt-key add -
-curl -s -L https://nvidia.github.io/nvidia-docker/$distribution/nvidia-docker.list | sudo tee /etc/apt/sources.list.d/nvidia-docker.list
+echo "========== Nvidia Docker Install =========="
+echo "Check Nvidia Docker Installed.."
+if type nvidia-docker &>/dev/null; then
+    echo "Docker Already Installed."
+else
+    echo "Installing Nvidia Docker..."
+    distribution=$(. /etc/os-release;echo $ID$VERSION_ID)
+    curl -s -L https://nvidia.github.io/nvidia-docker/gpgkey | sudo apt-key add -
+    curl -s -L https://nvidia.github.io/nvidia-docker/$distribution/nvidia-docker.list | sudo tee /etc/apt/sources.list.d/nvidia-docker.list
 
-sudo apt-get update && sudo apt-get install -y nvidia-container-toolkit
-sudo systemctl restart docker
+    sudo apt-get update && sudo apt-get install -y nvidia-container-toolkit
+    echo "Setting Nvidia Docker..."
+    sudo systemctl restart docker
+    echo "Now, Nvidia Docker Installed."
+fi
 
-# Download Typed Docker
+echo "========== Download Typed Docker =========="
 docker pull alstjr7375/typed
